+++ notificare-push/src/main/java/re/notifica/push/models/NotificareNotificationActionOpenedIntentResult.kt
@@ -1,0 +1,8 @@
+package re.notifica.push.models
+
+import re.notifica.models.NotificareNotification
+
+public data class NotificareNotificationActionOpenedIntentResult(
+    val notification: NotificareNotification,
+    val action: NotificareNotification.Action,
+)